+++ CHANGES.rst
@@ -12,6 +12,9 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
   [gbastien]
 - Fixed POD template `avis-df.odt` in `examples_fr` profile.
   [gbastien]
+- Make `CustomMeetingConfig.getUsedFinanceGroupIds` work with item sent to
+  another MC with inheritated advices.
+  [gbastien]
 
 4.2b24 (2022-09-29)
 -------------------